requirements.txt
google_generativeai
gspread==6.1.4
gspread_dataframe
numpy==1.26.4
pandas==2.2.3
selenium==4.26.1
webdriver_manager
pyyaml
plotly==6.0.0
yfinance==0.2.54
transformers
torch